replace existing tesselator
